AutoIdentify.mrc /Setpass passhere Will set the password on the network you used the command on, next time nickserv asks you to identify it will be done automatically.
AutoUnbanGetKeyInvite.mrc
Will try to use chanserv to auto unban, getkey, and invite yourself to the channel you tried to join if you recieve any of the error numerics. 473, 474 or 475.

CheckForBots.mrc
On join it checks for bots, if there are 2 or more bots it parts the channel, I made this for a bot I ran awhile ago

CommonAliases.mrc
Just various things, look through it for more details

PestControlWorldScript.mrc
Really old script I made, view the script for more information on what it does