Update 2 scripts to work with AutoIdentify.mrc
This updates AutoBypass.mrc and ConnectSetup.mrc.
If you have one or both loaded and have AutoIdentify.mrc loaded they
will work with it to try to provide more functionality.

AutoBypass will now try to identify you and reattempt a join on a
registered-user only channel.

ConnectSetup will try to authenticate you the second you connect to a
server.